test: refactor story assets
