+++ packages/ibm-products-web-components/src/components/side-panel/side-panel.stories.ts
@@ -14,11 +14,17 @@ import './index';
 // import Trashcan from '@carbon/icons/lib/trash-can/16';
 import { prefix } from '../../globals/settings';
 
-import styles from './story-styles.scss?lit';
-import { BUTTON_KIND } from '@carbon/web-components/es/components/button/defs.js';
 import '@carbon/web-components/es/components/button/index.js';
 import '@carbon/web-components/es/components/text-input/index.js';
 import '@carbon/web-components/es/components/textarea/index.js';
+
+import {
+  getContent,
+  getSubTitle,
+  getActionToolbarItems,
+  getActionItems,
+  getSlug,
+} from './_story-assets';
 
 const toggleButton = () => {
   document.querySelector(`${prefix}-side-panel`)?.toggleAttribute('open');
@@ -61,72 +67,6 @@ const contents = {
 
 const storyPrefix = 'side-panel-stories__';
 
-export const getContent = (index) => {
-  switch (index) {
-    case 1:
-      return html`
-        <style>
-          ${styles}
-        </style>
-        <h5>Section</h5>
-        <cds-text-input
-          label="Input A"
-          id="side-panel-story-text-input-a"
-          class="${storyPrefix}text-input"
-        ></cds-text-input>
-        <cds-text-input
-          label="Input B"
-          id="side-panel-story-text-input-b"
-          class="${storyPrefix}text-input"
-        ></cds-text-input>
-      `;
-    case 2:
-      return html`
-        <style>
-          ${styles}
-        </style>
-        <h5>Section</h5>
-        <div class="${storyPrefix}text-inputs">
-          <cds-text-input
-            label="Input A"
-            id="side-panel-story-text-input-a"
-          ></cds-text-input>
-          <cds-text-input
-            label="Input B"
-            id="side-panel-story-text-input-b"
-          ></cds-text-input>
-        </div>
-        <div class="${storyPrefix}text-inputs">
-          <cds-text-input
-            label="Input C"
-            id="side-panel-story-text-input-c"
-          ></cds-text-input>
-          <cds-text-input
-            label="Input D"
-            id="side-panel-story-text-input-d"
-          ></cds-text-input>
-        </div>
-        <div class="${storyPrefix}textarea-container">
-          <cds-textarea
-            label="Notes"
-            value="This is a text area"
-          ></cds-textarea>
-          <cds-textarea
-            label="Notes"
-            value="This is a text area"
-          ></cds-textarea>
-          <cds-textarea
-            label="Notes"
-            value="This is a text area"
-          ></cds-textarea>
-        </div>
-      `;
-
-    default:
-      return null;
-  }
-};
-
 const labels = {
   'No label': 0,
   'Shorter label': 1,
@@ -149,52 +89,10 @@ const subtitles = {
   'Short subtitle': 1,
   'Longer subtitle': 2,
 };
-export const getSubTitle = (index) => {
-  switch (index) {
-    case 1:
-      return html`<div slot="subtitle">This is your subtitle slot.</div>`;
-    case 2:
-      return html`<div slot="subtitle">
-        I am your subtitle slot for <strong>adding detail</strong> that can be
-        one or two lines.
-      </div>`;
-    default:
-      return null;
-  }
-};
 
 const actionToolbarItems = {
   'No action toolbar': 0,
   'With action toolbar': 1,
-};
-
-export const getActionToolbarItems = (index) => {
-  switch (index) {
-    case 1:
-      return html`
-        <cds-button slot="action-toolbar">Copy</cds-button>
-        <cds-button
-          slot="action-toolbar"
-          aria-label="Settings"
-          has-icon-only="true"
-          kind=${BUTTON_KIND.GHOST}
-          size="sm"
-          tooltip-text="Settings"
-        >
-        </cds-button>
-        <cds-button
-          slot="action-toolbar"
-          aria-label="Delete"
-          has-icon-only="true"
-          kind=${BUTTON_KIND.GHOST}
-          size="sm"
-          tooltip-text="Delete"
-        >
-        </cds-button>
-      `;
-    default:
-      return null;
-  }
 };
 
 const actionItems = {
@@ -207,98 +105,9 @@ const actionItems = {
   'Too many buttons': 6,
 };
 
-// TODO: There are problems switching this
-export const getActionItems = (index) => {
-  switch (index) {
-    case 1:
-      return html`<cds-button key="p" slot="actions" kind=${BUTTON_KIND.PRIMARY}
-        >Primary</cds-button
-      >`;
-    case 2:
-      return html`
-        <cds-button slot="actions" kind=${BUTTON_KIND.GHOST}>Ghost</cds-button>
-        <cds-button slot="actions" kind=${BUTTON_KIND.PRIMARY}
-          >Primary</cds-button
-        >
-      `;
-    case 3:
-      return html` <cds-button slot="actions" kind=${BUTTON_KIND.DANGER}
-          >Danger</cds-button
-        >
-        <cds-button slot="actions" kind=${BUTTON_KIND.PRIMARY}
-          >Primary</cds-button
-        >`;
-    case 4:
-      return html` <cds-button slot="actions" kind=${BUTTON_KIND.GHOST}
-          >Ghost</cds-button
-        >
-        <cds-button slot="actions" kind=${BUTTON_KIND.SECONDARY}
-          >Secondary</cds-button
-        >
-        <cds-button slot="actions" kind=${BUTTON_KIND.PRIMARY}
-          >Primary</cds-button
-        >`;
-    case 5:
-      return html`<cds-button
-          key="danger"
-          slot="actions"
-          kind=${BUTTON_KIND.DANGER}
-          >Danger</cds-button
-        >
-        <cds-button key="secondary" slot="actions" kind=${BUTTON_KIND.SECONDARY}
-          >Secondary</cds-button
-        >
-        <cds-button key="primary" slot="actions" kind=${BUTTON_KIND.PRIMARY}
-          >Primary</cds-button
-        >`;
-    case 6:
-      return html`<cds-button
-          key="danger"
-          slot="actions"
-          kind=${BUTTON_KIND.DANGER}
-          >Danger</cds-button
-        >
-        <cds-button key="tertiary" slot="actions" kind=${BUTTON_KIND.TERTIARY}
-          >Tertiary</cds-button
-        >
-        <cds-button key="secondary" slot="actions" kind=${BUTTON_KIND.SECONDARY}
-          >Secondary</cds-button
-        >
-        <cds-button key="primary" slot="actions" kind=${BUTTON_KIND.PRIMARY}
-          >Primary</cds-button
-        >`;
-    default:
-      return null;
-  }
-};
-
 const slugs = {
   'No Slug': 0,
   'With Slug': 1,
-};
-
-export const getSlug = (index) => {
-  switch (index) {
-    case 1:
-      return html`<cds-slug size="xs" alignment="bottom-right">
-        <div slot="body-text">
-          <p class="secondary">AI Explained</p>
-          <h1>84%</h1>
-          <p class="secondary bold">Confidence score</p>
-          <!-- //cspell: disable -->
-          <p class="secondary">
-            Lorem ipsum dolor sit amet, di os consectetur adipiscing elit, sed
-            do eiusmod tempor incididunt ut fsil labore et dolore magna aliqua.
-          </p>
-          <!-- //cspell: enable -->
-          <hr />
-          <p class="secondary">Model type</p>
-          <p class="bold">Foundation model</p>
-        </div>
-      </cds-slug>`;
-    default:
-      return null;
-  }
 };
 
 const defaultTemplate = {
